requirements.txt
streamlit
boto3
python-dotenv
pypdf2
faiss-cpu
numpy
